externalize kcp setup/teardown for ginkgo suite
